refactor: 컨포넌트 분리 #9
