+++ src/component/Switch/SwitchComponent.style.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const SwitchContainer = styled.div`
+    display: flex;
+    align-items: center;
+    padding: 1%;
+    position: fixed;
+    top: 0;
+    right: 0;
+`;
+
+export const SwitchText = styled.div`
+    font-weight: bold;
+    color: white;
+`;
